refactor(base-app): move ruler and mask to standalone feature modules
