+++ apps/base-app/src/app/contentScript/preloadFonts.ts
