+++ apps/base-app/src/features/readingToolsMask/config.ts
@@ -1,0 +1,2 @@
+export const BASE_COLOR = '#000000'
+export const BASE_MASK_HEIGHT = 30
