+++ apps/base-app/src/features/readingToolsRuler/config.ts
@@ -1,0 +1,2 @@
+export const BASE_COLOR = '#000000'
+export const BASE_RULE_SIZE = 1
