+++ apps/base-app/src/features/readingToolsRuler/index.ts
@@ -1,0 +1,1 @@
+export * as ReadingToolsRuler from './ui/ruler'
