+++ apps/base-app/src/shared/lib/dom.ts
@@ -1,0 +1,6 @@
+export const appendNewElementToBody = (id: string): HTMLElement => {
+  const divElement = document.createElement('div')
+  divElement.id = id
+  document.body.appendChild(divElement)
+  return divElement
+}
